refactor: player -> gamer로 package renaming
