+++ src/main/kotlin/blackjack/domain/card/PlayerCards.kt
@@ -1,6 +1,6 @@
 package blackjack.domain.card
 
-import blackjack.domain.player.PlayerName
+import blackjack.domain.gamer.PlayerName
 
 data class PlayerCards(
     val playerName: PlayerName,
