+++ src/main/kotlin/blackjack/domain/game/BlackJackGameTurn.kt
@@ -1,6 +1,6 @@
 package blackjack.domain.game
 
-import blackjack.domain.player.PlayerName
+import blackjack.domain.gamer.PlayerName
 
 sealed class BlackJackGameTurn {
