+++ src/main/kotlin/blackjack/domain/game/PlayerMatchResult.kt
@@ -1,6 +1,6 @@
 package blackjack.domain.game
 
-import blackjack.domain.player.PlayerName
+import blackjack.domain.gamer.PlayerName
 
 data class PlayerMatchResult(
     val playerName: PlayerName,
