+++ src/test/kotlin/blackjack/domain/game/CardDistributionResultTest.kt
@@ -3,7 +3,7 @@ package blackjack.domain.game
 import blackjack.domain.card.Card
 import blackjack.domain.card.CardHolder
 import blackjack.domain.card.PlayerCards
-import blackjack.domain.player.PlayerName
+import blackjack.domain.gamer.PlayerName
 import io.kotest.assertions.throwables.shouldNotThrow
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.StringSpec
