+++ src/test/kotlin/blackjack/domain/gamer/PlayerNames.kt
@@ -1,4 +1,4 @@
-package blackjack.domain.player
+package blackjack.domain.gamer
 
 fun playerNames(vararg names: String): PlayerNames {
     return PlayerNames(names.map { PlayerName(it) })
